Adding Math.Round()
The index calculated by snapPointY/itemHeight can be decimals sometimes, by adding Math.Round() we ensure that the index used in pickerData[index] is an integer and not float. 

When using a float in the index value, the selected value would result in "undefined" and would crash the app if used with the useState/useRef hook.  ( e.x pickerData[2.999]  == undefined) 

Also adding the same logic to the onFinish event, when onFinish is triggered the onEnd will not. As a result, the selected value would not be set.